refactor : getLastDay 함수 util 에 분리
